requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
jinja2>=3.1.3
alembic>=1.13.1
pydantic>=2.10.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
slowapi>=0.1.9
httpx>=0.27.0
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.3.0
mypy>=1.8.0
